src/PersonalCareerCoach/components/github_analyzer.py
from github import Github
import os

class GitHubAnalyzer:
    def __init__(self):
        self.github = Github(os.getenv("GITHUB_ACCESS_TOKEN"))

    def get_user_repos(self, username):
        user = self.github.get_user(username)
        return [repo.name for repo in user.get_repos() if not repo.fork]

    def analyze_repo(self, username, repo_name):
        repo = self.github.get_repo(f"{username}/{repo_name}")
        # Implement repo analysis logic here
        pass

    def get_most_relevant_project(self, username, job_description):
        repos = self.get_user_repos(username)
        # Implement project relevance scoring logic here
        pass