requirements.txt
streamlit
google-generativeai
python-dotenv
PyPDF2
requests
PyGithub
fastapi
uvicorn
python-multipart
-e .